first tests working for dmp
